Implemented API call
