+++ src/api/mentors-api.js
@@ -1,0 +1,5 @@
+import axios from 'axios';
+
+export default axios.create({
+    baseURL: 'https://codesydney-mentors.herokuapp.com/api/'
+});
